fix: remove benchmark_utils module again
