+++ datasets/libsvm.py
@@ -1,9 +1,10 @@
 from benchopt import BaseDataset, safe_import_context
 
 with safe_import_context() as import_ctx:
+    from scipy import sparse
+    from sklearn.feature_selection import VarianceThreshold
+    from sklearn.preprocessing import MaxAbsScaler, StandardScaler
     from libsvmdata import fetch_libsvm
-
-    from benchmark_utils import preprocess_data
 
 
 class Dataset(BaseDataset):
